Remove non-numba machinery
